+++ src/main/java/roomescape/booking/reservation/TossPaymentConfirmCommandFactory.java
@@ -1,7 +1,7 @@
 package roomescape.booking.reservation;
 
 import org.springframework.stereotype.Component;
-import roomescape.payment.dto.TossPaymentConfirmCommand;
+import roomescape.external.tosspayment.dto.TossPaymentConfirmCommand;
 
 @Component
 public class TossPaymentConfirmCommandFactory {
