+++ src/main/java/roomescape/config/RestClientConfiguration.java
@@ -5,8 +5,8 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.web.client.RestClient;
-import roomescape.payment.TossPaymentClient;
-import roomescape.payment.TossPaymentConfirmErrorHandler;
+import roomescape.external.tosspayment.TossPaymentClient;
+import roomescape.external.tosspayment.TossPaymentConfirmErrorHandler;
 
 @Configuration
 public class RestClientConfiguration {
