+++ src/main/java/roomescape/external/tosspayment/TossPaymentConfirmErrorHandler.java
@@ -1,4 +1,4 @@
-package roomescape.payment;
+package roomescape.external.tosspayment;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -7,7 +7,7 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.web.client.ResponseErrorHandler;
 import roomescape.exception.custom.reason.payment.PaymentException;
-import roomescape.payment.dto.TossErrorResponse;
+import roomescape.external.tosspayment.dto.TossErrorResponse;
 
 import java.io.IOException;
 import java.net.URI;
