+++ src/main/java/roomescape/external/tosspayment/dto/TossPaymentConfirmResponse.java
@@ -1,4 +1,4 @@
-package roomescape.payment.dto;
+package roomescape.external.tosspayment.dto;
 
 public record TossPaymentConfirmResponse(
         String paymentKey,
