+++ src/main/java/roomescape/reservationpayment/ReservationPaymentService.java
@@ -5,8 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.booking.reservation.TossPaymentConfirmCommandFactory;
-import roomescape.payment.TossPaymentAdapter;
-import roomescape.payment.dto.TossPaymentConfirmCommand;
+import roomescape.external.tosspayment.TossPaymentAdapter;
+import roomescape.external.tosspayment.dto.TossPaymentConfirmCommand;
 import roomescape.reservationpayment.dto.ReservationPaymentRequest;
 
 @Service
